Include the ability to force secure the app
When deployed to Heroku, we want to force the server to only listen for
serve via https. To do that, we use a middleware that will watch for
requests of type http and generate a redirect to the same page on https
instead.

This is done only for the heroku installed version, since otherwise we
would need to fiddle around with getting a local cert and that seems
like far too much work.