Resolve the awful login issue
It has taken a long time, but the ultimate issue with our login failing
seems to be related to the cookie being strict in relation to a
redirect of some sort?

This is not limited strictly to Chrome (but see the bug below) since it
also happens in FireFox, but when the cookie is strict, the browser
won't send it to us on a redirect or a fetch, which means a hard reload
is needed.

By setting the cookie type to Lax we get the cookie sent to us, but it
won't get sent in fetches, which is cool because we don't need to be
hammered with it every time anyway.

https://bugs.chromium.org/p/chromium/issues/detail?id=696204