Disable pre-rendering during site builds
According to the very fine manual *cough* *cough* we should not be doing
any pre-rendering if the page content is not expected to be the same
for any two users.

Since our page is customized per user (or most are anyway), that seems
like a good reason to actually turn this off. This is handy because
we're also triggering code when we do this that perhaps best not be
triggered.