Upgrade to the latest heroku stack
We're upgrading to the latest heroku stack, and best practices seem to
indicate that we should include the expected stack version in the
app.json file.

For us that's not currently needed, but it might be in the future and
why borrow problems when you can nip them right in the butt?