Update our installation of TailwindCSS
In looking up why we were getting strange errors from yarn during
deployment it transpires that possibly the instructions on the
SvelteKit page for using Tailwind with SvelteKit were out of date,
since the official docs mention more packages than that which we had
previously installed.

This includes the extras, and as added measure alters the setup to match
the TailwindCSS docs for this instead, which in some theoretical way
are more up to date most likely (except for how SvelteKit is a moving
target). For example, the SvelteKit documentation assumed TailwindCSS
version 2.