Set a default npm version in package engines
Include a default npm version in the package.json to ensure that when
heroku uses it (if it does) it uses an expected version.

We're not doing anything with it since we use yarn, but other tools
might and without an explicit version being specified it's possible
that the default could change and alter behaviour, which we don't
want.