Include a simple dark mode toggle
Tailwind supports a dark mode by default, which works by allowing you to
specify what properties need to change as a result of dark mode, so
we're using that here.

In particular we are using the "class" variant; by default it wants to
use dark mode if the OS prefers it, but in the interests of serving
people whose OS does not support that properly or at all, as well as
for those that would rather make the decision per app, we allow
toggling that.