Allow the dark mode button to appear on small size
The button for toggling between dark and light mode was set as:
    hidden md:block

this makes it be hidden, but re-enabled on medium and larger, which has
the effect of the button being invisible on phones. This was an
inadvertent slip since the original button was for showing
notifications, which would not be displayed on a small screen in the
same way.

This resolves the issue, though the button positioning is slightly wrong
in this case (a little too far to the right).