Get ready for our server side code push
We're working on a set of changes by which our back end code will be
able to work with Prisma databases to persist data as well as talk to
Twitch.

To that end we're setting up a new entry point for such setup code which
will be triggered by the load of the hooks file, which happens lazily
on the first request after startup.

This is not perfect and we may try to fix it later, but it's good enough
for now.