Include an initial Prisma schema for our project
This initial version of things sets us up with a simple Schema that can
track Twitch users and (in theory) their tokens, although for the
forseeable future we're not interested in actually persisting those
(it will be used to track the token for the user that represents the
bot).

As a part of this change we're also populating the list of known Twitch
users in the front page instead of the packed in data.