Include encryption logic in our Database Code
We will soon be storing tokens in the database, and so we want to make
sure that we encrypt them for extra safety, both to make sure we don't
bone up something live on stream, as well as to provide protection for
other leaks.

This is using the same encryption code that we used in a prior project.