Set up a chat back end for the chat bot component
This sets up our initial chat system for the integration service.

In a nutshell, part of the configuration is the userId of the person
that represents the service (as a bot), and when that user logs in for
the first time, their token is kept and used to connect to chat.

At startup, an existing token is looked for, and if found, will be used
to connect to chat right away.

Some pruning is also done so that if there is saved information about a
bot token and the configured userId of the bot changes, we can redact
that information away.

This is smart enough to connect to Twitch chat right away at startup if
the token is provided, though due to the vagaries of SvelteKit it won't
happen until the first request is made, which is mildy sad.