Include new models and seed data for our addons
The ultimate goal of this project is to provide a simple system by which
non-developers can play with our "addons" or "extensions" for Twitch
streams without having to run their own bot and know coding.

Here we're adding in models to track such addons and their information
as well as the records for users actually applying them to their
channel.

This includes also a data seeing mechanism to put the correct data into
the database at startup if it's not already there.