Replace the front page user cards with addon cards
As a step towards something that is actually useful as a tool, this
change replaces the front page user cards with cards that show you what
the currently available addons in the system are.

This needs more work because the cards should have more information in
them, including links to go to their page and also to add them to your
profile if you're logged in, but that will come in future updates.