Update the profile page to include used addons
This augments the GET method of the addons API to allow it to filter the
list of addons by those that are in use by a particular user.

With this in place, this update also brings to the profile page the list
of addons that the user currently has configured, if they're logged in.

Still missing is the ability to actually add or remove one without using
database tools, as well as visualizing the pages for the addons
themselves.