Include a page slug in the addon data schema
In order to make links nicer, it's probably a good idea that there be a
unique string that we can use in a URL that isn't a randomly generated
(and this mildly tricky to remember) identifier.

For this purpose we're implementing a page slug, which can be used in
the url as a more human readable way to address addons.