Refactor the auth token cookie handling code
Since we're going to be using the authorization cookie to gate API
access to things, this change pulls the code that used to be inline in
the back end handling code out into its own reusable module.

This still looks up and verifies the token, but returns it directly or
returns null, depending on whether there was a cookie and if it's valid.

Currently we don't distinguish between there being no auth cookie and
there being a cookie with an invalid token.