Remove the unused user endpoint and card component
The initial prototypes of the page as we were learning about SvelteKit
and how its endpoint system works used the list of currently known
users as data points.

Now that we have actual addons, we don't need that endpoint (and if we
did it's easy enough to add back), these parts can be removed for
clarity's sake.