Refactor some of the endpoints for getting addons
In the previous incarnation there was a single API endpoint for getting
the list of addons, which could optionally be filtered by the addons
in use by a specific user (which was used in the profile page).

Under this new design there's a single endpoint to get either all of the
adds or the details on a particular one based on it's ID and/or slug,a
as well as a new user data endpoint that will return all of the addons
for the currently logged in user.

The access here is implicit; it requires the cookie that provides the
authorization JWT to know who's making the request and thus what data
to return.