Include a simple addon page route
This needs some serious work, but it puts in the underpinnings to being
able to navigate to different addon pages from the main page and from
the user's profile page.

This modifies the AddonCard component so that the name of the component
that it display is actually a link based on the addon's slug, which
allows the new page to be navigated to.