Include extra yarn scripts
This simple change includes a few things that allow us to perform some
tasks easier.

The first is a database setup script that we can execute that will
update the schema in the database and force potential data loss if
there are issues. This is intended to be used only interactively after
setting the database URL to the one used by the actual application
itself, but is here to remind us how to do that.

The second is a simple alias that will allow showing the papertail web
console for the application, which allows for nicer logging.