Fix insecure site redirection
We had some code in place which was intended to forward an incoming
request using the http protocol to an https address instead, to ensure
that the application always uses SSL, since Heroku does not enforce
this externally.

It turns out however that regardless of the protocol used on the
incoming request, the Heroku router sends us the request as if it was
using the https protocol, so this code never fired.

This fixes that by checking to see if the header that Heroku includes to
indicate what the forwarded protocol was is insecure, instead of the
protocol in the URL. 

This is likely not restricted purely to Heroku since according to their
docs there are a variety of packages that can do this for you at the
express level.