Alter the application name
Now that we're closer to having an actual stub base application, this
changes the name of the application to be more appropriate.

As a part of this change, the Heroku app was renamed, along with the
name of the repo, various git remotes, and the link to the application
that is contained in its own footer.