Fix a dependabot issue by removing ksuid-cli
This is a development only package, which would be stripped out of any
releases (Heroku drops devDependencies before building the slug, for
example), and as such the security issue related to an older version is
not strictly required to fix. Still, that banner is the spooky.

This removes the dependency to fix the security issue is brings; a
locally installed version will be used instead.