Switch to the latest version of SvelteKit
This updates our SvelteKit install (plus TailwindCSS and Svelte itself)
to their latest versions; SvelteKit is currently under active
development and they're not afraid to make breaking changes in version
1 code, so it seems like a good idea to keep abreast of updates as they
happen.

This does **NOT** update the node-adapter because going to the next
point revision seems to break the app for seemingly unknown reasons;
the changelog only mentions that the version of TypeScript changed
(despite what appears to be 200+ files changing, many of which are not
typescript).

We should probably be wary about updating that one in the future.