Upgrade SvelteKit, adapter-node and vite
A couple of days ago we attempted to update things so as to keep
ourselves up to date with the goings-on in SvelteKit, since it is a
moving target, but ran afoul of a problem in adapter-node which was
causing us issues.

Since then, several updates have occurred to both packages, and some
playing in a side-side project also shows that vite can be upgraded
safely as well

This change alters our versions of node and kit to be the appropriate
versions, while also making sure that we version lock ourselves at the
"next" tag for the time being.