Clean up how we import our Svelte components
This is a small change, but brings some nice readability. There are now
indexes in the components directory that re-export all of the
components, which allows us to import any component with a more simple
import statement.

To make this look nicer, the vite config was modified to change what the
alias for the components directory is named, so it reads more like
English.