Start our work on including DaisyUI for our UI
After some playing around, I've decided that I like the DaisyUI project,
which contains a fair few out of the box components that are styled via
Tailwind, and includes a superior theme engine as well as more compound
tailwind properties so that our markup doesn't look as horrible.

This change includes only the library and makes the modification needed
so that the theme will change between our two themes when clicked; other
things don't look right yet, but that's coming.