Implement a new page footer using DaisyUI
This is similar to the previous one, but uses a poop icon instead of
text like the old one did, and also includes a link to the YouTube
channel as well.

As a part of this change, the layout of the page in general was changed
to not be padded on all edges.