Update our Navigation bar to use DaisyUI
This is a rather large structural change to things, but one which makes
the actual files more succinct and easier to read. This modifies the
navigation bar at the top of the page so that it uses DaisyUI, which is
less complicated in terms of how the menu is handled.