Update our seeding script and seed data
In order to help with testing and mocking things up, we're including an
third addon that contains the union of the available properties.

In addition, since we expect these things to be changing fairly rapidly,
we're also modifying the seeding script so that the upsert will update
the data in the database when we run it, if anything changes.

Since there's no in-app way to update the addon entries and they're
expected to be seeded in, it should be fairly safe to do this since the
only way we might clobber data is if we manually edit it instead of
editing the seed.

So, point me at this commit when I do something dumb.