Update the addon service and component card
We want to make sure that the addons are returned in the order in which
they were created, and for that we can search based on the ksuid, which
encodes the timestamp of the creation of the key id.

In addition to that, this adds in the missing badge display for whether
a component has chat integrations, overlay integrations, or both.