Update our AddonCard implementation and API
To flesh out our Addon cards a bit more, this updates the back end API
so that when requesting the list of addons a new 'installed' field is
set to `true` or `false` to indicate if that addon is installed by the
current user or not.

In cases where the request is made but there's not a user logged in,
this field is always `false`.

The short card component for addons now uses this information to
determine if the button on the card should be for adding or removing
the addon; when there is no user logged in, the button is not present.