Fix an issue with user specific addon queries
In the previous commit we augmented the API that returns addon
information to include a field that indicates if that addon is in use
for the currently logged in user or not (if there is one).

This fixes an issue with that, in that we forgot that the query to
get ONLY the addons for a user needs to have the value as well (where it
will always be true).