Alter the Addon Schema as relates to the configs
Previously our schema was set up so that there would be a URL for
fetching the configuration schema and a second URL for fetching default
config options for a new install of the addon.

This has been changed so that the configuration information is inlined
directly in the configuration object itself instead of being fetchable
via a URL, and the schema also contains the default values which
obviates the need for a default item as well.

The seed data has also been updated to populate in schema information.

The queries for addon data have been modified so that when they return
the record, the configuration object is parsed into a JSON object for
return instead of being returned as a string.