Include a baseDir option to get the root path
This includes a new configuration item named baseDir which contains the
full absolute path to the root of the application's installation
folder.

Although it's not currently needed, it might be in the future and we
currently have the code handy from a previous test scenario for how we
are setting up our schemas, so we're including it here in case we need
it later.