Update the Addon schema to rename the title
The field in the table named title was actually meant to be a short
description of the addon for use in the addons page, but we ended up
accidentally using it as a page title because the name wasn't very
good.

This changes the name of the field in the schema to track that.

In addition, the AddonCard has been modified, and now includes an extra
prop that allows you to specify which of the long or short descriptions
to use.