Implement some simple test components for configs
Include some simple components for working with the different types of
configuration fields that addons might have: strings, numbers, booleans
and enums.

These are still a work in progress but this sets up the basic structure
for how these should work and be deployed based on the individual
configurations of each of the addons.

Currently they use the default values for each item, though they need
to actually be bound to the configs in use for a particular user.

As a part of this change, the packed in data has been augmented as well.