Create initial port of the back end code
This is the initial version of the first port of the code that used to
be the API section of the routes in the SvelteKit Ruinous Addons
application.

All of that endpoints for that have been moved out of the UI portion and
altered to work using Express. 

This includes support for our web sockets and also sets up the
appropriate back end code for ensuring that the Twitch code gets setup
at application launch and not during the first request.