Clean up our API routing methods
This changes up the routing so that all of the routes are applied in one
place instead of in separate modules, which gives a better sense for
the full structure of the available routes.

In addition, we also include some catch-all routes so that we can
generate appropriate errors on invalid requests.