Include our overlay redirection system
One extra portion of the back end that we're moving from the other
project into this one is the system by which a request for an overlay
with a specific ID gets translated into a redirection to the actual
overlay page itself.

This doesn't include the code that builds the static overlays, but it
DOES include everything else.

As a side note, it's unclear how well this will work in the production
environment, since the routing to get to us is different and express is
a bit of a dill weed about giving you an easy way to get the URL, and I
don't trust myself to have done it correctly...