Bundle in our sample overlays from the UI project
This last change puts the building blocks in place to be able to do our
overlay development.

We set up vite to build the overlays (and also serve them as needed),
and also modified the sample code that was stubbed out in the UI so
that it can use the socket server code to talk back and forth with the
server.