Alter how we add our routes to the server
This change modifies the code that adds the routes that drive the API
and the other back end endpoints so that instead of adding the routes
directly it creates and returns a middleware Router that can be used in
the main code.

This is partially driven by hoping that we can get the routes for this
to play nice with the Vite dev server, but even if not this is still a
cleanup to the code.