Swap our application back end framework
After some pondering and more work on the UI end, it has been decided
that the best course of action is to re-combine the API and UI portions
into a single project, but using rollup instead of vite so that we
don't have issues with running multiple servers.

This commit is the framework for all of that, and includes some
temporary pages and components just as a proof of concept; these need
to be fleshed out into something more concrete from our other project.