Include new API endpoint for getting the user
Previously, SvelteKit had a server side handler we needed to implement
that would pipeline our core back end data directly into the front end
side of the application, in particular the information on the current
user.

This is now our responsibility, so as a first step this includes the API
needed for this, which is a new endpoint which uses the provided auth
token cookie to either return back information on the user to which it
belongs, or an empty object if the user is not logged in.