Hook up a store for user information for logins
This is a small change, but it includes a new store that contains the
information on the currently logged in user, which is fetched when a
back end page load happens and is available everywhere on the front end.

All of the front end code has been altered to use this information to
know about the currently logged in user, including if there is one at
all.