Re-implement the previous profile page
This is a first pass at a new implementation of the profile page; it
behaves as the old one did, but could be made better through the use of
a store to keep track of the addons that the user has, since that
really only needs to be set and adjusted occasionally and not on every
page load.

That aspect is being left pending the implementation of the core addons
slug pages, which will be the final part of the initial port.