Add in the dedicated per-addon page slug handler
This brings in a route for /addons/:slug, which takes the user to a page
that is dedicated to a specific addon, providing information about its
configuration information and other details.

This is based on the previous iteration and does not yet handle errors
correctly, but it's functional nonetheless as a first draft.