Implement a simple SPA router change on the server
This is a small change, but since as of right now we only need to serve
a single static file manually in response to a client side route doing
a hard reload, having a generic function for this does not make a ton
of sense.

In addition it's handy to see what the route being requested actually is
since that can help with debugging, so we're including that as a log
as well.

To accomplish this we also have a new runtime configuration option for
where the static content is going to be served from.