Implement the insecure-to-secure redirect code
In the previous iteration we detected when the incoming forwarded
request was not secure, and forced the browser to redirect to a secure
version.

This implements the same thing, but via a library instead of as a
SvelteKit hook, due to a distinct lack of SvelteKit, which makes the
hook problematic.