Include some simple winston logging
This implements some simple console-only log transports for being able
to log in a way that makes it easier to see when all logs happened,
their severity, and where they came from.

When deployed this would allow us to control better what the contents of
the logs look like.

Currently this is console only because we're using papertrail on the
heroku end, which provides all of the file based and other searchable
backing we could want. Logging to postgres seems rather heavyweight on
our local end for information that's largely just diagnostic in nature.