Update the login route so that it can redirect
This alters the route and handler for /login so that if there is any
extra path on the end, the site will redirect to that location on a
success.

This is currently only implemented for the back end; the front end code
has not been adjusted to know that this is possible yet.