Introduce nodejs aliases for the server code
This change is an experiment with NodeJS subpath import patterns, which
allow us to set up import aliases that make relative imports a thing of
the past.

We're settled on some names for the core bits of the server code, and
the list will undoubtedly be subject to change as we go forward and
find loopholes or change names to be clearer.