Refactor the seeding data layout to make life good
The upcoming changes are going to need to be able to pull information
about a specific addon, and it is needed at compile time rather than at
runtime.

This restructures the seed data structure to make it easier to get at
the information for specific addons from both the server side and
client side versions of the code, which both need to share the same
information.