Implement the core structure for addon support
In order to function the application needs server side code that is
associated with each individual addon. This change sets that up by
having a new set of functions that are used to initialize the code for
each addon.

The addon overlays are also fleshed out a bit so that they can get
information on the overlay they're for, which allows them to talk to
the appropriate back end code for the addon that the overlay is for.

The actual structure of what happens on the server and client side is
still TBD, and will be fleshed out shortly.