Fix the terrible circular dependency issue
This resolves the issues we were having with circular dependencies in
the components subtree, which was caused by having components import the
index of components when they load (which is of course dumb).

Now all components do relative imports for any siblings they need, and
the index files only export those components which are expected to be
used from outside of the page component section (i.e. in page endpoints
and the like).