Fix the issue with the login redirection
A previous change we made allowed the /login route to take an additional
path and then redirect to there once the login succeeds.

This works, but the method for determining the path to redirect to was
naive and did not take hashes or query strings into account. This rears
it's ugly head when you try to force login to change the user, since
that includes extra URL in the form of a query string.

We now parse the URL to get just the path name, so that this cannot
happen.

This is kind of ham handed at the moment and could probably be done
better; we're jamming a fake protocol on there just to make the URL
parser happy.