Fix the theme being forced to dark mode
During early development of the rewrite and before we got the nav bar
ported over from the original version, the theme was hard coded at the
document level to be dark.

This redacts that change, since it, you know, forces the theme to always
be dark. As nice as it would be to force people that like light mode to
see how annoying it is to not have a toggle, we're going to be the
bigger man here.