Fix user data loading race condition
While trying to implement the profile page sending the user to the login
page if not logged in, we uncovered a race condition in which the page
can fully load before the request for user data comes back.

This change encompasses ensuring that the load of the user data is
completed prior to allowing any portion of the page to load. Using
this, we can rely on the fact that $user is the appropriate data
everywhere.

As a part of this change, we also completed the profile page redirect
change, included a new routine for fetching JSON data since it will be
used in a couple of places, and include a spinner courtesy of axel669
(author of svelte-doric) for while the request is pending.

The load routine can enforce a delay while in development mode to give
an indication of where loads are happening that might be lengthy.

There were also a couple of places where links needed to have the
appropriate use: added to them so that they would be client side routed,
and the explicit logout button in the profile page was force linked to
localhost from a previous debug session, which is now fixed.