Update prisma and our seed data
Update to the latest version of Prisma, 4.3.1.

In addition, the sample seed data for addons was modified so that they
follow a unique naming scheme for their data, as they previously also
did for their addonId values.

This will allow future changes to import with boilerplate without having
to know actual symbol names; only the files will differ.