Alter our configuration to need only one root URL
In the past we had two distinct ways to specify how to get at the
overlay files, because the front end and back end ran on different
servers during testing (and for a while were also separate projects
too).

That is no longer the case, and having two of them are redundant, since
we now need a third more generic one.

This change removes the overlay specific options in favor of just the
base that the application is served from, allowing us to simulate the
others as needed.