Fix the FOUC for overlay pages
The overlay pages have some global CSS that was injected via JS code at
page load time; this allows for a brief FOUC as the page renders, then
gets the appropriate CSS and reflows itself.

Here we're fixing that by setting up the rollup config to have postcss
extract the CSS to a file so that it can be loaded more conventionally.