Make Lighthouse happy w/r/t SEO
This is a simple one, but apparently if you tell a web crawler that it
can crawl your web, it buys you cake and takes you out to dinner and
maybe a movie, but if you don't, various web tools take a dump on your
head?

The life of a web developer is hard indeed.

In any case, this includes a robots.txt file as well as a top level
metatag with a site description, so that in case the site ever
accidentally gets crawled, google will know what to do.