Improve accessibility for blind buttons and links
There were a few buttons and links which link only to an icon and don't
contain text which needed some labels attached to them to help with
accessibility.

This applies those changes so that things are more consistently
discoverable for screen readers and the like, though more work needs to
be done overall on this.