First pass at using Svelte for the overlays
This sets up the boilerplate code that will serve as our code base as we
work through the best way to finalize this change without having any
redundant files.

What we have now is an Overlay component whose job it is to ensure that
it fetches the overlay information it needs from the back end and then,
when it does, renders a component that represents the overlay.

This puts all of the logic directly into the Svelte file, and includes
a custom rollup build plugin that allows us to generate all of the
overlays without having to have multiple copies of the boilerplate code
in place.