Allow overlays to get at their own sockets
As a bit of a blind spot in the initial refactoring of the overlay code
we neglected to have any way for the common startup code to convey to
the overlay it's socket handle so that it can do communications.

This change alters the contract a little bit so that when the launch
happens, the overlays are given not only the details of their overlay
data but also the socket handle, should they be desirous of socket
action.