Allow the bot to dynamically join and leave chat
This provides an update to the endpoints for installing and uninstalling
addons so that when either operation is taken, the bot will either join
or leave the channel as appropriate.

For a join, the bot knows to join the channel if an addon is installed
for a user and it's local chat client is not already in that channel.

For a part, the bot knows to leave the channel if, after uninstalling
an addon, there are no remaining addons for this user that require chat
access.