Include support for CORS handling
For at least testing of things, we need the ability to talk to the
socket server from domains other than those from which the application
is served.

This implements a simple configuration that allows for specifying a
list of 0 or more hosts to be added to the CORS origin setting in
the server.