Include a socket manager class
The back end code needs to be able to keep a list of all of the incoming
connection, including who they belong to and what addon they're for.

This adds such a set of routines, which, given a socket and information
about it, will populate a table in a way that allows you to look up
the information for a socket or for a particular combination of userId
and addonId.