Implement our connection API for overlays
This fleshes out our handling for connections of websockets from
overlays.

In particular, when a socket connects it will transmit a message asking
for the information on the overlay to which it represents, and in
response to that will announce itself with a HELO message to let the
back end know what overlay it represents.