Switch to MongoDB from PostgreSQL
Since the Heroku free tier of Postgres is going away, we need to migrate
the project to something that is more affordable than their offering.
Some quick research shows that MongoDB Atlas has a decent free getting
started tier and what appears to be fair pricing based on our needs.

This set of changes migrates the existing application from using psql
to mongodb:

  * Upgraded the version of Prisma in use (unrelated, but fun anyway)
  * Update the schema; tables have an id now since Mongo requires that
  * Schema now uses Json instead of String for JSON data
  * Seeding data uses ksuid for new id values
  * Seeding data now uses bare JSON instead of strings for config infos
  * Query code no longer needs to parse JSON since it's already present
  * OverlayId is now sent back from the list of installed addons