Fix our chat join/part code
The original implementation has a bug in which it checks the chat client
existence using the wrong sentinel value, which causes any install and
uninstall of an addon with a chat component to fail, possibly crashing
out the server.

Sadly this was not noticed until after the initial change was deployed.