Set up our initial testing routines
There is still much work to be done here, and the methodology is not
at 100% yet, but this sets up some tests using aegis by @axel669 to
allow us to test our API endpoints.

This currently assumes that the local server is running and hits the
server live in order to validate the API. This could/should be changed
to mock up the server using the route endpoint functions directly or
similar so that we can do testing during deployments and pushes, etc.