Include schema testing routines
This augments the tests from the previous commit so that they can also
validate that the data coming out of the endpoints follows the correct
schema, for which we are using the `joker` library by @axel669.

There is a bit more work to be done here, in that the configSchema is
not validated currently, being as the schema of each object is distinct
based on its type.