Include new user endpoint for config manipulation
There was not previously an API for fetching the configuration of an
addon, nor was there a way to adjust it.

This commit addresses both, including a new API endpoint which allows
directly fetching and setting the configuration information for a
specific addon, and a test suite to help verify the thing.

This also includes a new error handler for handling cases where the body
of an incoming request has a JSON content type but is not valid JSON.
Previously this would immediately bail the request with a 400 error,
now we handle that more gracefully by way of a middleware being
applied.