Clean up and refactor our API test method
This replaces the previous `api` and `apiJSON` functions with a new
`request` and `requestWithAuth` function.

The previous functions were for getting either just the response object
or the object and also the body, but this required us to re-wrap it in
every test file to get custom behaviour, like specifying other options.

The new `request` function is all encompassing in this regard, with
defaults to fit the most common cases for clarity, and the
`requestWithAuth` variant is a wrapper that makes it less verbose to
specify auth tokens.