Augment the config schema for enum items
Previously enumerations were just an array of strings; these are now an
array of key-value pairs that specify the textual human name and the
shorter computer value.

In addition the visualization uses those now, and also uses a
placeholder as an initial disabled item as a reminder for what the
value is for.