Implement testing configuration schema validation
This implements validation for the configuration schema in the addon
information. This is done in a way that verifies that for each type of
value that can appear in a configuration item that the schema that
describes it is correct.

This is used only for the testing process and validates the actual
schema itself. The actual configuration items will be done next.