Implement actual config schema validation
This adds in a new set of routines that can validate that a potential
configuration item that is to be used to set up configuration (or to
validate a retrieved config, should you care).

This is used in the testing code to verify that configurations are
valid, though much more work needs to be done on this (but we are
changing the way errors present).