Remove interim field from addon results
While gathering the list of addons, we temporarily inject a timestamp
into the object so we can order them by creation date.

This doesn't need to be exposed outwardly (at least for now), so for
cleanliness we redact it away before return and after it's used to
perform the sort.